Modificaciones en producto y principal con ordenes de trabajo detalladas

+++ GestionIntegral/CapaDatos/Pedido.cs
@@ -9,7 +9,7 @@ namespace GestionIntegral.CapaDatos
     {
         int idDetallePedido;
         int idPedido;
-        Double total;
+        Decimal total;
 
         DateTime fecha;
         int idCliente ;
@@ -29,10 +29,10 @@ namespace GestionIntegral.CapaDatos
         public string NumGuia { get => numGuia; set => numGuia = value; }
         public string MedioPago { get => medioPago; set => medioPago = value; }
         public DateTime? FechaPago { get => fechaPago; set => fechaPago = value; }
-        public Double Total { get => total; set => total = value; }
+        public Decimal Total { get => total; set => total = value; }
         public DateTime? FechaEnvio { get => fechaEnvio; set => fechaEnvio = value; }
 
-        public Pedido(int idDetallePedido, DateTime fecha, int idCliente, int idEstado, string numGuia, string medioPago, DateTime? fechaPago, float total, DateTime? fechaEnvio)
+        public Pedido(int idDetallePedido, DateTime fecha, int idCliente, int idEstado, string numGuia, string medioPago, DateTime? fechaPago, Decimal total, DateTime? fechaEnvio)
         {
             IdDetallePedido = idDetallePedido;
             Fecha = fecha;
